# backend/openapi.yaml
openapi: 3.1.0
info:
  title: 補助金申請書自動生成API
  version: 1.0.0
  description: 補助金申請書類の自動生成、検証、エクスポートを行うAPI
  contact:
    name: API Support
    email: support@subsidy-app.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-api-version-info:
    deprecated: false
    sunset: null
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.subsidy-app.example.com
    description: Production server
...
paths:
...
  /intake/upload:
    post:
      tags:
        - intake
      summary: ファイルアップロード
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - type
              properties:
                file:
                  type: string
                  format: binary
                  description: アップロードファイル（PDF、画像等）
                type:
                  type: string
                  enum: [certificate, tax_return, other]
                  description: 書類種別
      responses:
        '200':
          description: アップロード成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntakeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
  /intake/extract:
    post:
      tags:
        - intake
      summary: データ抽出
      operationId: extractData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_id
              properties:
                file_id:
                  type: string
                  description: アップロードファイルID
                ocr_provider:
                  type: string
                  enum: [tesseract, cloud]
                  default: tesseract
      responses:
        '200':
          description: 抽出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  schemas:
...
    IntakeResponse:
      type: object
      properties:
        file_id:
          type: string
        sha256:
          type: string
        mime_type:
          type: string
        size:
          type: integer
        status:
          type: string
          enum: [uploaded, scanning, ready, rejected]
    ExtractResponse:
      type: object
      properties:
        extraction_id:
          type: string
        file_id:
          type: string
        document_type:
          type: string
          enum: [certificate, tax_return_personal, tax_return_corporate, unknown]
        extracted_fields:
          type: object
          additionalProperties: true
        confidence_scores:
          type: object
          additionalProperties:
            type: number
            minimum: 0
            maximum: 1
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: リクエスト不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: ファイルサイズ超過
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []