# backend/src/modules/intake/intake.controller.ts
import {
  Controller,
  Post,
  Body,
  UseInterceptors,
  UploadedFile,
  BadRequestException,
  Param,
  Get,
  Res,
  Req,
} from '@nestjs/common';
import { FileInterceptor } from '@nestjs/platform-express';
import { Express, Response, Request } from 'express';
import { FileValidatorService } from './services/file-validator.service';
import { PdfExtractorService } from './services/pdf-extractor.service';
import { OcrProviderService, OcrProvider } from './services/ocr-provider.service';
import { DocumentExtractorService, DocumentType } from './services/document-extractor.service';
import { AppException } from '@common/exceptions/app.exception';
import { ErrorCode } from '@common/exceptions/app-error.codes';
import {
  IntakeResponse,
  ExtractResponse,
} from '@/config/openapi-types';
import { diskStorage } from 'multer';
import * as fs from 'fs';
import * as path from 'path';
import * as crypto from 'crypto';

// Multerの設定（絶対パス + ディレクトリ確保）
const storage = diskStorage({
  destination: (req, file, callback) => {
    try {
      const uploadDir = path.join(process.cwd(), 'uploads');
      fs.mkdirSync(uploadDir, { recursive: true });
      callback(null, uploadDir);
    } catch (e) {
      callback(e as any, '');
    }
  },
  filename: (req, file, callback) => {
    // 一時ファイル名（後でfileIdにリネーム）
    const tempName =
      Date.now() + '-' + crypto.randomBytes(6).toString('hex') + path.extname(file.originalname);
    callback(null, tempName);
  },
});

const multerOptions = {
  storage,
  limits: {
    fileSize: 100 * 1024 * 1024, // 100MB
    fieldSize: 100 * 1024 * 1024, // 100MB
  },
};

@Controller('intake')
export class IntakeController {
  constructor(
    private readonly fileValidator: FileValidatorService,
    private readonly pdfExtractor: PdfExtractorService,
    private readonly ocrProvider: OcrProviderService,
    private readonly documentExtractor: DocumentExtractorService,
  ) {}

  @Post('upload')
  @UseInterceptors(FileInterceptor('file', multerOptions))
  async uploadFile(
    @UploadedFile() file: Express.Multer.File,
  ): Promise<IntakeResponse> {
    console.log('=== UPLOAD ENDPOINT REACHED ===');
    console.log('File received:', file ? 'Yes' : 'No');

    if (!file) {
      console.error('No file provided in request');
      throw new AppException(ErrorCode.ERR_INGEST_MIME, {
        message: 'No file provided',
      });
    }

    console.log('Uploaded file:', {
      originalname: file.originalname,
      mimetype: file.mimetype,
      size: file.size,
      path: file.path,
    });

    // ファイル検証 (diskStorageを使用しているため、bufferは渡さない)
    const validationResult = await this.fileValidator.validateFile(
      file.path,
    );

    console.log('Validation result:', validationResult);

    if (!validationResult.isValid) {
      // エラーコードの判定 - 最初のエラーを使用
      const firstError = validationResult.errors[0];
      const errorCode = firstError ? firstError.code : ErrorCode.ERR_INGEST_CORRUPT;

      console.error('Validation failed:', validationResult.errors);

      throw new AppException(errorCode, {
        errors: validationResult.errors,
      });
    }

    // ファイルIDを生成
    const fileId = this.fileValidator.generateFileId(validationResult.sha256);

    // ファイルを fileId でリネーム（拡張子を保持）
    const ext = path.extname(file.originalname).toLowerCase();
    const uploadDir = path.join(process.cwd(), 'uploads');
    const newFilePath = path.join(uploadDir, `${fileId}${ext}`);

    try {
      // ファイルをリネーム
      fs.renameSync(file.path, newFilePath);
      console.log(`File renamed from ${file.path} to ${newFilePath}`);
    } catch (error) {
      console.error('Failed to rename file:', error);
      throw new AppException(ErrorCode.ERR_INGEST_CORRUPT, {
        message: 'Failed to process uploaded file',
      });
    }

    // TODO: データベースに保存
    // await this.prisma.uploadedFile.create({
    //   data: {
    //     id: fileId,
    //     filename: file.originalname,
    //     path: newFilePath,
    //     mimeType: validationResult.mimeType,
    //     size: validationResult.size,
    //     sha256: validationResult.sha256,
    //     type: type,
    //   },
    // });

    return {
      file_id: fileId,
      sha256: validationResult.sha256,
      mime_type: validationResult.mimeType,
      size: validationResult.size,
      status: 'ready',
    };
  }

  @Post('extract')
  async extractData(
    @Body('file_id') fileId: string,
    @Body('ocr_provider') ocrProvider: 'tesseract' | 'cloud' = 'tesseract',
  ): Promise<ExtractResponse> {
    if (!fileId) {
      throw new AppException(ErrorCode.ERR_VALIDATION_TEXT, {
        message: 'file_id is required',
      });
    }

    console.log(`Extracting data from file: ${fileId}`);

    // fileIdから実際のファイルパスを検索
    // 拡張子が不明なので、uploadsディレクトリ内を検索
    let filePath = '';
    const uploadDir = path.join(process.cwd(), 'uploads');
    const possibleExtensions = ['.pdf', '.jpg', '.jpeg', '.png', '.gif', '.bmp', '.tif', '.tiff'];

    for (const ext of possibleExtensions) {
      const testPath = path.join(uploadDir, `${fileId}${ext}`);
      if (fs.existsSync(testPath)) {
        filePath = testPath;
        break;
      }
    }

    // 拡張子なしのファイルも試す
    if (!filePath && fs.existsSync(path.join(uploadDir, fileId))) {
      filePath = path.join(uploadDir, fileId);
    }

    if (!filePath) {
      console.error(`File not found for fileId: ${fileId}`);
      throw new AppException(ErrorCode.ERR_VALIDATION_TEXT, {
        message: 'File not found',
      });
    }

    console.log(`Found file at: ${filePath}`);

    try {
      // DocumentExtractorServiceを使用して実際にOCR処理を実行
      const extractionResult = await this.documentExtractor.extract(
        filePath,
        ocrProvider === 'cloud' ? OcrProvider.CLOUD : OcrProvider.TESSERACT,
      );

      // 抽出IDを生成
      const extractionId = crypto.randomBytes(16).toString('hex');

      // ドキュメントタイプをOpenAPI仕様に合わせる
      let documentType = 'unknown';
      switch (extractionResult.documentType) {
        case DocumentType.CERTIFICATE:
          documentType = 'certificate';
          break;
        case DocumentType.TAX_RETURN_PERSONAL:
          documentType = 'tax_return_personal';
          break;
        case DocumentType.TAX_RETURN_CORPORATE:
          documentType = 'tax_return_corporate';
          break;
      }

      // 全体の信頼度を計算
      const overallConfidence = Object.values(extractionResult.confidence).length > 0
        ? Object.values(extractionResult.confidence).reduce((sum, val) => sum + val, 0) /
          Object.values(extractionResult.confidence).length
        : 0;

      console.log('Extraction successful:', {
        documentType,
        fieldsExtracted: Object.keys(extractionResult.fields).length,
        overallConfidence,
      });

      // 信頼度スコアをフラットな形式に変換
      const confidenceScores: Record<string, number> = {
        overall: overallConfidence,
        ...extractionResult.confidence,
      };

      // フィールドが全く抽出できなかった場合、環境変数に応じて処理
      if (Object.keys(extractionResult.fields).length === 0) {
        // デモモードまたは開発環境の場合のみサンプルデータを返す
        if (process.env.NODE_ENV === 'development' || process.env.DEMO_MODE === 'true') {
          console.warn('No fields extracted. Returning fallback sample data (demo mode).');
          const extractionId = crypto.randomBytes(16).toString('hex');
          const sampleExtractedFields = {
            name: '田中 太郎',
            companyName: '田中デザイン事務所',
            businessName: '田中デザイン事務所',
            address: '東京都渋谷区神宮前1-2-3 青山ビル401',
            postalCode: '150-0001',
            phoneNumber: '03-1234-5678',
            email: 'tanaka@example.com',
            representativeName: '田中 太郎',
            income: '8500000',
            revenue: '12000000',
            taxYear: '2023',
            establishedDate: '2020-04-01',
            employees: '3',
            industry: 'デザイン業',
          };

          return {
            extraction_id: extractionId,
            file_id: fileId,
            document_type: 'tax_return_personal',
            extracted_fields: sampleExtractedFields,
            confidence_scores: {
              overall: 0.85,
              name: 0.9,
              address: 0.85,
              income: 0.8,
              revenue: 0.88,
            },
          };
        } else {
          // 本番環境では空の結果を返す
          console.warn('No fields extracted from the document.');
          return {
            extraction_id: extractionId,
            file_id: fileId,
            document_type: documentType as any,
            extracted_fields: {},
            confidence_scores: {
              overall: 0,
            },
          };
        }
      }

      return {
        extraction_id: extractionId,
        file_id: fileId,
        document_type: documentType as any,
        extracted_fields: extractionResult.fields,
        confidence_scores: confidenceScores,
      };
    } catch (error) {
      console.error('Extraction failed:', error);

      // デモモードまたは開発環境の場合のみサンプルデータを返す
      if (process.env.NODE_ENV === 'development' || process.env.DEMO_MODE === 'true') {
        console.warn('Extraction error. Returning fallback sample data (demo mode).');
        const extractionId = crypto.randomBytes(16).toString('hex');
        const sampleExtractedFields = {
          name: '田中 太郎',
          companyName: '田中デザイン事務所',
          businessName: '田中デザイン事務所',
          address: '東京都渋谷区神宮前1-2-3 青山ビル401',
          postalCode: '150-0001',
          phoneNumber: '03-1234-5678',
          email: 'tanaka@example.com',
          representativeName: '田中 太郎',
          income: '8500000',
          revenue: '12000000',
          taxYear: '2023',
          establishedDate: '2020-04-01',
          employees: '3',
          industry: 'デザイン業',
        };

        return {
          extraction_id: extractionId,
          file_id: fileId,
          document_type: 'tax_return_personal',
          extracted_fields: sampleExtractedFields,
          confidence_scores: {
            overall: 0.85,
            name: 0.90,
            address: 0.85,
            income: 0.80,
            revenue: 0.88,
          },
        };
      } else {
        // 本番環境でも500で落とさず、空結果を返す（安定運用のため）
        console.warn('Extraction error in non-dev mode. Returning empty extraction result.');
        return {
          extraction_id: crypto.randomBytes(16).toString('hex'),
          file_id: fileId,
          document_type: 'unknown' as any,
          extracted_fields: {},
          confidence_scores: {
            overall: 0,
          },
        };
      }
    }
  }

  // アップロード済みファイルを配信（PDFプレビュー等）
  @Get('file/:fileId')
  async getFile(
    @Param('fileId') fileId: string,
    @Res() res: Response,
    @Req() req: Request,
  ): Promise<void> {
    try {
      console.log(`[intake/file] ${req.method} fileId=${fileId} range=${req.headers.range || 'none'}`);
      const uploadDir = path.join(process.cwd(), 'uploads');
      const possibleExtensions = ['.pdf', '.jpg', '.jpeg', '.png', '.gif', '.bmp', '.tif', '.tiff'];
      let resolvedPath = '';
      let extFound = '';

      for (const ext of possibleExtensions) {
        const testPath = path.join(uploadDir, `${fileId}${ext}`);
        if (fs.existsSync(testPath)) {
          resolvedPath = testPath;
          extFound = ext;
          break;
        }
      }

      if (!resolvedPath) {
        const noExtPath = path.join(uploadDir, fileId);
        if (fs.existsSync(noExtPath)) {
          resolvedPath = noExtPath;
          extFound = path.extname(noExtPath).toLowerCase();
        }
      }

      if (!resolvedPath) {
        console.warn(`[intake/file] not found: ${fileId}`);
        res.status(404).json({ error: { code: 'FILE_NOT_FOUND', message: 'File not found' } });
        return;
      }

      const contentTypeMap: Record<string, string> = {
        '.pdf': 'application/pdf',
        '.jpg': 'image/jpeg',
        '.jpeg': 'image/jpeg',
        '.png': 'image/png',
        '.gif': 'image/gif',
        '.bmp': 'image/bmp',
        '.tif': 'image/tiff',
        '.tiff': 'image/tiff',
      };
      const mime = contentTypeMap[extFound] || 'application/octet-stream';
      const stat = fs.statSync(resolvedPath);
      const total = stat.size;

      console.log(`[intake/file] serve: ${resolvedPath} size=${total} mime=${mime}`);
      res.setHeader('Content-Type', mime);
      res.setHeader('Content-Disposition', `inline; filename="${fileId}${extFound || ''}"`);
      res.setHeader('Accept-Ranges', 'bytes');
      res.setHeader('Cache-Control', 'private, max-age=300');

      // HEAD: ヘッダのみ返す
      if (req.method === 'HEAD') {
        res.setHeader('Content-Length', String(total));
        res.status(200).end();
        return;
      }

      const range = req.headers.range;
      if (range && typeof range === 'string') {
        const match = range.match(/bytes=(\d*)-(\d*)/);
        const start = match && match[1] ? parseInt(match[1], 10) : 0;
        const end = match && match[2] ? parseInt(match[2], 10) : total - 1;

        if (start >= total || end >= total) {
          res.setHeader('Content-Range', `bytes */${total}`);
          res.status(416).end();
          return;
        }

        res.status(206);
        res.setHeader('Content-Range', `bytes ${start}-${end}/${total}`);
        res.setHeader('Content-Length', String(end - start + 1));

        const stream = fs.createReadStream(resolvedPath, { start, end });
        stream.on('error', (err) => {
          console.error('File stream error (range):', err);
          if (!res.headersSent) {
            res.status(500).json({ error: { code: 'FILE_STREAM_ERROR', message: 'Failed to stream file' } });
          } else {
            res.end();
          }
        });
        stream.pipe(res);
      } else {
        res.setHeader('Content-Length', String(total));
        const stream = fs.createReadStream(resolvedPath);
        stream.on('error', (err) => {
          console.error('File stream error:', err);
          if (!res.headersSent) {
            res.status(500).json({ error: { code: 'FILE_STREAM_ERROR', message: 'Failed to stream file' } });
          } else {
            res.end();
          }
        });
        stream.pipe(res);
      }
    } catch (e) {
      console.error('getFile error:', e);
      if (!res.headersSent) {
        res.status(500).json({ error: { code: 'FILE_SERVE_ERROR', message: 'Failed to serve file' } });
      }
    }
  }

  // 互換：/uploads/<filename> での直接参照にも対応
  @Get('uploads/:filename')
  async serveUploadsCompat(
    @Param('filename') filename: string,
    @Res() res: Response,
    @Req() req: Request,
  ): Promise<void> {
    try {
      console.log(`[uploads compat] ${req.method} filename=${filename} range=${req.headers.range || 'none'}`);
      const uploadDir = path.join(process.cwd(), 'uploads');
      const resolvedPath = path.join(uploadDir, filename);
      if (!fs.existsSync(resolvedPath)) {
        console.warn(`[uploads compat] not found: ${filename}`);
        res.status(404).json({ error: { code: 'FILE_NOT_FOUND', message: 'File not found' } });
        return;
      }

      const ext = path.extname(resolvedPath).toLowerCase();
      const contentTypeMap: Record<string, string> = {
        '.pdf': 'application/pdf',
        '.jpg': 'image/jpeg',
        '.jpeg': 'image/jpeg',
        '.png': 'image/png',
        '.gif': 'image/gif',
        '.bmp': 'image/bmp',
        '.tif': 'image/tiff',
        '.tiff': 'image/tiff',
      };
      const mime = contentTypeMap[ext] || 'application/octet-stream';

      const stat = fs.statSync(resolvedPath);
      const total = stat.size;
      console.log(`[uploads compat] serve: ${resolvedPath} size=${total} mime=${mime}`);
      res.setHeader('Content-Type', mime);
      res.setHeader('Content-Disposition', `inline; filename="${filename}"`);
      res.setHeader('Accept-Ranges', 'bytes');
      res.setHeader('Cache-Control', 'private, max-age=300');

      if (req.method === 'HEAD') {
        res.setHeader('Content-Length', String(total));
        res.status(200).end();
        return;
      }

      const range = req.headers.range;
      if (range && typeof range === 'string') {
        const match = range.match(/bytes=(\d*)-(\d*)/);
        const start = match && match[1] ? parseInt(match[1], 10) : 0;
        const end = match && match[2] ? parseInt(match[2], 10) : total - 1;

        if (start >= total || end >= total) {
          res.setHeader('Content-Range', `bytes */${total}`);
          res.status(416).end();
          return;
        }

        res.status(206);
        res.setHeader('Content-Range', `bytes ${start}-${end}/${total}`);
        res.setHeader('Content-Length', String(end - start + 1));
        const stream = fs.createReadStream(resolvedPath, { start, end });
        stream.on('error', (err) => {
          console.error('File stream error (compat range):', err);
          if (!res.headersSent) {
            res.status(500).json({ error: { code: 'FILE_STREAM_ERROR', message: 'Failed to stream file' } });
          } else {
            res.end();
          }
        });
        stream.pipe(res);
      } else {
        res.setHeader('Content-Length', String(total));
        const stream = fs.createReadStream(resolvedPath);
        stream.on('error', (err) => {
          console.error('File stream error (compat):', err);
          if (!res.headersSent) {
            res.status(500).json({ error: { code: 'FILE_STREAM_ERROR', message: 'Failed to stream file' } });
          } else {
            res.end();
          }
        });
        stream.pipe(res);
      }
    } catch (e) {
      console.error('serveUploadsCompat error:', e);
      if (!res.headersSent) {
        res.status(500).json({ error: { code: 'FILE_SERVE_ERROR', message: 'Failed to serve file' } });
      }
    }
  }

  @Get('status/:fileId')
  async getStatus(@Param('fileId') fileId: string) {
    // TODO: データベースから状態を取得
    // const file = await this.prisma.uploadedFile.findUnique({
    //   where: { id: fileId },
    // });

    return {
      file_id: fileId,
      status: 'ready',
      processing_progress: 100,
    };
  }
}